refactored SuperTypesLastSorter to a static utility class
