+++ src/main/java/org/mockito/internal/util/reflection/SuperTypesLastSorter.java
@@ -17,11 +17,15 @@ import java.util.List;
  * moved after their subclasses.
  */
 public class SuperTypesLastSorter {
+    
+    private SuperTypesLastSorter() {
+    }
+    
     /**
      * Return a new collection with the fields sorted first by name,
      * then with any fields moved after their supertypes.
      */
-    public List<Field> sort(Collection<? extends Field> unsortedFields) {
+    public static List<Field> sortSuperTypesLast(Collection<? extends Field> unsortedFields) {
         List<Field> fields = new ArrayList<Field>(unsortedFields);
 
         Collections.sort(fields, compareFieldsByName);
@@ -53,6 +57,7 @@ public class SuperTypesLastSorter {
 
 
     private static final Comparator<Field> compareFieldsByName = new Comparator<Field>() {
+        @Override
         public int compare(Field o1, Field o2) {
             return o1.getName().compareTo(o2.getName());
         }
